+++ src/common/PersonTile/styled.js
